Selector.Option -> SelectorOption
