+++ packages/snaps-sdk/src/jsx/components/form/SelectorOption.test.tsx
@@ -1,0 +1,16 @@
+import { Option } from './Option';
+
+describe('Option', () => {
+  it('renders a dropdown option', () => {
+    const result = <Option value="foo">Foo</Option>;
+
+    expect(result).toStrictEqual({
+      type: 'Option',
+      props: {
+        value: 'foo',
+        children: 'Foo',
+      },
+      key: null,
+    });
+  });
+});
