+++ packages/snaps-sdk/src/jsx/components/form/SelectorOption.ts
@@ -1,0 +1,44 @@
+import { createSnapComponent } from '../../component';
+import type { CardElement } from '../Card';
+
+/**
+ * The props of the {@link SelectorOption} component.
+ *
+ * @property value - The value of the selector option. This is used to populate the
+ * state in the form data.
+ * @property children - The component to display.
+ */
+export type SelectorOptionProps = {
+  value: string;
+  children: CardElement;
+};
+
+const TYPE = 'SelectorOption';
+
+/**
+ * A selector option component, which is used to create a selector option. This component
+ * can only be used as a child of the {@link Selector} component.
+ *
+ * @param props - The props of the component.
+ * @param props.value - The value of the selector option. This is used to populate the
+ * state in the form data.
+ * @param props.children - The component to display.
+ * @returns A selector option element.
+ * @example
+ * <Selector name="selector">
+ *  <Selector.Option value="option1"><Card title="Option 1" value="Foo" /></Selector.Option>
+ *  <Selector.Option value="option2"><Card title="Option 2" value="Bar" /></Selector.Option>
+ *  <Selector.Option value="option3"><Card title="Option 3" value="Baz" /></Selector.Option>
+ * </Selector>
+ */
+export const SelectorOption = createSnapComponent<
+  SelectorOptionProps,
+  typeof TYPE
+>(TYPE);
+
+/**
+ * A selector option element.
+ *
+ * @see SelectorOption
+ */
+export type SelectorOptionElement = ReturnType<typeof SelectorOption>;
