+++ packages/snaps-sdk/src/jsx/components/form/index.ts
@@ -8,7 +8,8 @@ import type { InputElement } from './Input';
 import type { OptionElement } from './Option';
 import type { RadioElement } from './Radio';
 import type { RadioGroupElement } from './RadioGroup';
-import type { SelectorElement, SelectorOptionElement } from './Selector';
+import type { SelectorElement } from './Selector';
+import type { SelectorOptionElement } from './SelectorOption';
 
 export * from './Button';
 export * from './Checkbox';
@@ -21,6 +22,7 @@ export * from './FileInput';
 export * from './Form';
 export * from './Input';
 export * from './Selector';
+export * from './SelectorOption';
 
 export type StandardFormElement =
   | ButtonElement
